Python/microRobotFramework_python_04/requirements.txt
pyserial>=3.5
numpy>=1.19.0
typing>=3.7.4
dataclasses>=0.6; python_version<"3.7"
